+++ speech/__init__.py
@@ -1,0 +1,142 @@
+#!/usr/bin/env python
+"""
+
+    Greynir: Natural language processing for Icelandic
+
+    Copyright (C) 2021 Miðeind ehf.
+
+       This program is free software: you can redistribute it and/or modify
+       it under the terms of the GNU General Public License as published by
+       the Free Software Foundation, either version 3 of the License, or
+       (at your option) any later version.
+       This program is distributed in the hope that it will be useful,
+       but WITHOUT ANY WARRANTY; without even the implied warranty of
+       MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+       GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with this program.  If not, see http://www.gnu.org/licenses/.
+
+
+    Icelandic text to speech via TTS web services.
+
+    Includes friendly command line interface.
+
+"""
+
+from typing import List, Iterable, Dict, Set
+from types import ModuleType
+
+import os
+import sys
+import logging
+import importlib
+
+
+DEFAULT_VOICE = "Dora"
+VOICES_DIR = "voices"
+
+# Text formats
+# For details about SSML markup, see:
+# https://developer.amazon.com/en-US/docs/alexa/custom-skills/speech-synthesis-markup-language-ssml-reference.html
+DEFAULT_TEXT_FORMAT = "ssml"
+SUPPORTED_TEXT_FORMATS = frozenset(("text", "ssml"))
+assert DEFAULT_TEXT_FORMAT in SUPPORTED_TEXT_FORMATS
+
+# Audio formats
+DEFAULT_AUDIO_FORMAT = "mp3"
+SUPPORTED_AUDIO_FORMATS = frozenset(("mp3", "ogg_vorbis", "pcm"))
+assert DEFAULT_AUDIO_FORMAT in SUPPORTED_AUDIO_FORMATS
+
+
+def load_voice_modules() -> Dict[str, ModuleType]:
+    """Dynamically load all voice modules, map voice ID
+    strings to the relevant modules."""
+
+    prevdir = os.getcwd()
+    abspath = os.path.abspath(__file__)
+    dname = os.path.dirname(abspath)
+    os.chdir(dname)
+    sys.path.append(dname)
+
+    def modules_in_dir(directory: str) -> List[str]:
+        """Find all python modules in a given directory"""
+        files = os.listdir(directory)
+        modnames: List[str] = list()
+        for fname in files:
+            if not fname.endswith(".py"):
+                continue
+            if fname.startswith("_"):  # Skip any files starting with _
+                continue
+            mod = directory.replace("/", ".") + "." + fname[:-3]  # Cut off .py
+            modnames.append(mod)
+        return modnames
+
+    v2m = {}
+    for modname in modules_in_dir(VOICES_DIR):
+        try:
+            # Try to import
+            m = importlib.import_module(modname)
+            if not m:
+                continue
+            voices: Iterable[str] = getattr(m, "VOICES")
+            if not voices:
+                continue  # No voices declared, skip
+            for v in voices:
+                v2m[v] = m
+        except Exception as e:
+            print(f"Error importing voice module {modname}: {e}")
+
+    os.chdir(prevdir)
+
+    return v2m
+
+
+VOICE_TO_MODULE = load_voice_modules()
+SUPPORTED_VOICES = VOICE_TO_MODULE.keys()
+
+
+def _sanitize_args(args: Dict) -> Dict:
+    """Make sure arguments to speech synthesis functions are sane."""
+    # Make sure we have a valid voice ID
+    voice_id = args["voice_id"].lower().capitalize()
+    if voice_id not in SUPPORTED_VOICES:
+        logging.warning(
+            f"Voice '{voice_id}' not in supported voices, reverting to default ({DEFAULT_VOICE})"
+        )
+        args["voice_id"] = DEFAULT_VOICE
+    else:
+        args["voice_id"] = voice_id
+
+    # Clamp speed to 50-150% range
+    args["speed"] = max(min(1.5, args["speed"]), 0.5)
+
+    return args
+
+
+def text_to_audio_data(
+    text: str,
+    text_format: str = DEFAULT_TEXT_FORMAT,
+    audio_format: str = DEFAULT_AUDIO_FORMAT,
+    voice_id: str = DEFAULT_VOICE,
+    speed: float = 1.0,
+) -> bytes:
+    """Returns audio data for speech-synthesised text."""
+    args = locals().copy()
+    module = VOICE_TO_MODULE.get(voice_id)
+    fn = getattr(module, "text_to_audio_data")
+    return fn(**_sanitize_args(args))
+
+
+def text_to_audio_url(
+    text: str,
+    text_format: str = DEFAULT_TEXT_FORMAT,
+    audio_format: str = DEFAULT_AUDIO_FORMAT,
+    voice_id: str = DEFAULT_VOICE,
+    speed: float = 1.0,
+) -> str:
+    """Returns URL to audio of speech-synthesised text."""
+    args = locals().copy()
+    module = VOICE_TO_MODULE.get(voice_id)
+    fn = getattr(module, "text_to_audio_url")
+    return fn(**_sanitize_args(args))
